+++ src/Sinks/SinkMobitech.php
@@ -21,9 +21,10 @@ class SinkMobitech extends SinkBase
     public function destinationTables(): array
     {
         return [
-            'mobitech_actors'       => 'Actor/operator names with ID from other Mobitech tables',
-            'mobitech_statistics'   => 'Passenger statistics from Mobitech/ORCA',
-            'mobitech_transactions' => 'Transactions from Mobitech/ORCA',
+            'mobitech_actors' => 'Actor/operator names with ID from other Mobitech tables',
+            'mobitech_softpay_transactions' => 'Softpay transactions from Mobitech',
+            'mobitech_statistics' => 'Passenger statistics from Mobitech',
+            'mobitech_transactions' => 'AutoPass transactions from Mobitech/ORCA',
         ];
     }